+++ CHANGELOG.md
@@ -3,6 +3,9 @@ https://github.com/josdejong/mathjs
 
 ## <not yet released>, version 0.4.1
 
+- Changed: changed behavior of the power operator in the parser from left
+  associative to right associative in order to follow the mathematical
+  conventions.
 - Changed: renamed the parsers method 'put' into 'set'.
 - Fixed: method 'in' did not check for units to have the same base.